# Conflicts:
#	src/routers/review.ts

+++ src/routers/review.ts
@@ -7,7 +7,7 @@ import Stores from "../schemas/store"
 const reviewRouter = express.Router();
 
 reviewRouter.post("/", authMiddleware, async (req, res) => {
-  const { userId } = res.locals.user;
+  const userId = res.locals.user._id;
   const { orderId, content, star } = req.body;
   const date = new Date();
   try {
@@ -18,10 +18,9 @@ reviewRouter.post("/", authMiddleware, async (req, res) => {
     let menuIdList : Array<string> = []
     for (let i = 0; i<isExistCheckReview.menus.length; i++){
       menuIdList.push(isExistCheckReview.menus[i].menuId)
-      console.log(isExistCheckReview.menus[i])
-      console.log(isExistCheckReview.menus[i].menuId)     
     }
     
+    console.log("userId:", userId);
 
     const currentReview = await Reviews.create({
       storeId : isExistCheckReview.storeId,
@@ -52,7 +51,7 @@ reviewRouter.post("/", authMiddleware, async (req, res) => {
 
     res.status(200).json({ message: "success", reviewId: currentReview._id });
   } catch (err) {
-    res.status(400).json({ message: "fail", err: err });
+    res.json({ message: "fail", err: err });
   }
 });
 
@@ -61,40 +60,58 @@ reviewRouter.get("/:storeId", async (req, res) => {
   try {
     const isReviewExist = await Reviews.find({ storeId: storeId }).exec();
     if (isReviewExist.length == 0) {
-      return res.status(400).json({ message: "리뷰가 존재하지 않습니다." });
+      return res.json({ message: "리뷰가 존재하지 않습니다." });
     }
     res.status(200).json({ message: "success", review: isReviewExist });
   } catch (err) {
-    res.status(400).json({ message: "fail", err: err });
+    res.json({ message: "fail", err: err });
   }
 });
 
-reviewRouter.put("/:reviewId", async (req, res) => {
+reviewRouter.put("/:reviewId", authMiddleware, async (req, res) => {
   const { reviewId } = req.params;
   const { content, star } = req.body;
 
   try {
+    const review = await Reviews.findOne({ _id: reviewId });
+    const userId = review.userId
+
+    if (String(userId) != String(res.locals.user._id)) {
+      res.json({ message: "wrong user" });
+
+      return;
+    }
+
     const changedReview = await Reviews.findOneAndUpdate(
       { _id: reviewId },
       { content, star }
     ).exec();
     res.status(200).json({ message: "success", review: changedReview });
   } catch (err) {
-    res.status(400).json({
+    res.json({
       message: "fail",
       err: err,
     });
   }
 });
 
-reviewRouter.delete("/:reviewId", async (req, res) => {
+reviewRouter.delete("/:reviewId", authMiddleware, async (req, res) => {
   const { reviewId } = req.params;
+
+  const review = await Reviews.findOne({ _id: reviewId });
+  const userId = review.userId
+
+  if (String(userId) != String(res.locals.user._id)) {
+    res.json({ message: "wrong user" });
+
+    return;
+  }
 
   try {
     await Reviews.deleteOne({ _id: reviewId }).exec();
     res.status(200).json({ message: "success" });
   } catch (err) {
-    res.status(400).json({
+    res.json({
       message: "fail",
       err: err,
     });
